Address EVP flagevaluation review feedback
