+++ utils/_features.py
@@ -2709,9 +2709,9 @@ class _Features:
     def feature_flags_evp_flagevaluation(test_object):
         """Feature Flags EVP Flagevaluation
 
-        https://feature-parity.us1.prod.dog/#/?feature=548
+        https://feature-parity.us1.prod.dog/#/?feature=540
         """
-        return _mark_test_object(test_object, feature_id=548, owner=_Owner.ffe)
+        return _mark_test_object(test_object, feature_id=540, owner=_Owner.ffe)
 
     @staticmethod
     def feature_flags_event_enrichment(test_object):
